% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CCC}
\alias{CCC}
\title{The CCC Algorithm}
\usage{
CCC(matrix, overlap = 0.1)
}
\arguments{
\item{matrix}{The input matrix in which the columns are ordered by time.}

\item{overlap}{Numeric in ]0,1[] containing the maximum percentage of overlapping allowed. The \code{info} slot will contain a vector which filters the discovered Biclusters based on this value.}
}
\value{
A Biclust S4 Class object containing extra information of the CCC algorithm result in the \code{info} slot.
}
\description{
A R-wrapper which directly calls the original Java code for the CCC algorithm (\url{http://kdbio.inesc-id.pt/software/ccc-biclustering/}) and transforms it to the output format of the \code{Biclust} R package.
}
\details{
PLACEHOLDER
}
\examples{
\dontrun{
data(heatdata)
out <- CCC(heatdata)
info(out,method="bonferroni")
}
}
\author{
Ewoud De Troyer
}
\references{
Sara C. Madeira, Miguel C. Teixeira, Isabel Sa Correia and Arlindo L. Oliveira, "Identification of Regulatory Modules in Time Series Gene Expression Data using a Linear Time Biclustering Algorithms", \emph{IEEE/ACM Transactions on Computational Biology and Bioinformatics}
}

