% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CCC_disc}
\alias{CCC_disc}
\title{Discretization used for CCC algorithm.}
\usage{
CCC_disc(matrix, threshold = 1)
}
\arguments{
\item{matrix}{Gene Expression Matrix}

\item{threshold}{Threshold for second step in discretization procedure.}
}
\value{
Returns a discretized matrix (see Description)
}
\description{
Discretization procedure proposed by Ji and Tan (see References). A \code{M x N} input matrix results in a \code{M x (N-1)} discretized matrix.
}
\author{
Ewoud De Troyer
}
\references{
L. Ji and K. Tan. Mining gene expression data for positive and negative co-regulated gene clusters. Bioinformatics, 20(16):2711-2718, 2004.

L. Ji and K. Tan. Identifying time-lagged gene clusters using gene expression data. Bioinformatics, 21(4):509-516, 2005.
}

