% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCCPackage.R
\docType{package}
\name{CCCBiclust}
\alias{CCCBiclust}
\alias{CCCBiclust-package}
\title{Identification of Regulatory Modules in Time Series Gene Expression Data through CCC and e-CCC biclustering algorithm.}
\description{
CCCBiclust is an R wrapper for the available Java code for the CCC and e-CCC algorithm available at \url{http://kdbio.inesc-id.pt/software/ccc-biclustering/} and \url{http://kdbio.inesc-id.pt/software/e-ccc-biclustering/}
}
\references{
Sara C. Madeira, Miguel C. Teixeira, Isabel Sa Correia and Arlindo L. Oliveira, "Identification of Regulatory Modules in Time Series Gene Expression Data using a Linear Time Biclustering Algorithms", \emph{IEEE/ACM Transactions on Computational Biology and Bioinformatics}

Sara C. Madeira and Arlindo L. Oliveira, "A polynomial time biclustering algorithm for finding genes with approximate expression patterns in gene expression time series", \emph{Algorithms for Molecular Biology 2009}, 4:8 (4 June 2009).
}

