% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eCCC}
\alias{eCCC}
\title{The e-CCC Algorithm}
\usage{
eCCC(matrix, minr = 1, minc = 1, maxErrors = 1, overlap = 0.1)
}
\arguments{
\item{matrix}{The input matrix in which the columns are ordered by time.}

\item{minr}{Integer containing the row quorum (= minimum number of genes allowed in e-CCC-Biclusters).}

\item{minc}{Integer containing the column quorum (= minimum number of contiguous time points allowed in e-CCC-Biclusters).}

\item{maxErrors}{Integer containing the amount of errors allowed, per gene, in the e-CCC-Biclustering algorithm (value of e).}

\item{overlap}{Numeric in ]0,1[ containing the maximum percentage of overlapping allowed. The \code{info} slot will contain a logical vector which can filter the biclusters which overlap more than \code{overlap*100} percent.}
}
\value{
A Biclust S4 Class object containing extra information of the e-CCC algorithm result in the \code{info} slot.
}
\description{
A R-wrapper which directly calls the original Java code for the e-CCC algorithm (\url{http://kdbio.inesc-id.pt/software/e-ccc-biclustering/}) and transforms it to the output format of the \code{Biclust} R package.
}
\details{
PLACEHOLDER
}
\examples{
\dontrun{
data(heatdata)
out <- eCCC(heatdata,minr=3,minc=2) 
eCCCinfo(out,filter="Bonf0.01")
out@info$eCCCpatterns$BC53[1:10,]
}
}
\author{
Ewoud De Troyer
}
\references{
Sara C. Madeira and Arlindo L. Oliveira, "A polynomial time biclustering algorithm for finding genes with approximate expression patterns in gene expression time series", Algorithms for Molecular Biology 2009, 4:8 (4 June 2009)
}

