% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{eCCCinfo}
\alias{eCCCinfo}
\title{Filter eCCCinfo in info slot of eCCC Result}
\usage{
eCCCinfo(resbic, filter = "none")
}
\arguments{
\item{resbic}{Result from eCCC.}

\item{filter}{Choose one of the following filters: \code{"none", "Bonf0.01","Bonf0.01_overlap"}. Respectively no biclusters are filtered, significant biclusters with under 1 percent after Bonferroni correction are filtered and biclusters after Bonferroni (1 percent) with taking the maximum overlap into account are filtered.}
}
\value{
Returns a filtered and ordered eCCCinfo (info slot of eCCC result)
}
\description{
Reorders eCCCinfo slot based on the p-values order and applies one of the 2 available filters.
}
\examples{
\dontrun{
data(heatdata)
out <- eCCC_ext(heatdata,minr=3,minc=2)
eCCCinfo(out,filter="Bonf0.01_overlap")
}
}
\author{
Ewoud De Troyer
}

