% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{heatdata}
\alias{heatdata}
\title{Heat Stress Data}
\format{A matrix with 5811 genes and 5 conditions}
\description{
Gene expression data related with the yeast response to heat stress (see Reference).
}
\references{
A. P. Gasch, P. T. Spellman, C. M. Kao, O. Carmel-Harel, M. B. Eisen, G. Storz, D. Votstein, and P. O. Brown. Genomic expression programs in the response of yeast cells to environmental changes. Molecular Biology of the Cell, 11:4241-4257,2000.
}

